API-аудит: суточный лимит сообщений и сверка фактов в документации (#282)

+++ sdk/go/examples/upload.go
@@ -86,7 +86,7 @@ func main() {
 					Key:      key,
 					Name:     filename,
 					FileType: pachca.FileTypeFile,
-					Size:     int32(fileSize),
+					Size:     int64(fileSize),
 				},
 			},
 		},
